Remove trace constraint from trace spec, and led TLC model checker with `IsEvent` predicate.
This predicate could check if event name is given (or not) in `desc` field of an event. It reduces considerably the statespace by removing non-deterministic aspects that comes from disjunction of actions in `Next`. For example:

`Next = ActionA \/ ActionB \/ ActionC`, TLC have to search if log event match with `ActionA`, `ActionB` or `ActionC`. If we precise the event name when logging, TLC doesn't have to make this sort of search.

It will be also possible to precise (or not) some parameter values of an action that match with a log event in order to reduce non-deterministic due to the existential quantifiers. For example, `\E i \in Server : BecomeLeader(i)` could be replaced by BecomeLeader(logline.node) if node is given when logging.